backend-microservice/auth/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.2.0
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.3.0
ecdsa==0.19.0
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.1.0
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.1
pyotp==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-jose==3.3.0
qrcode==7.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
